Для ролей, freelance и client созданы useCase для проверки доступов. Для всех UseCase из Auth добавлена проверка на доступ. Тесты - с ошибками

+++ app/Modules/Auth/Application/Actions/Client/RemoveClientUseCase.php
@@ -5,6 +5,8 @@ namespace App\Modules\Auth\Application\Actions\Client;
 use App\Modules\Auth\Application\Interfaces\ClientRepositoryInterface;
 use App\Modules\Auth\Application\Interfaces\StaffRepositoryInterface;
 use App\Modules\Auth\Infrastructure\Models\Staff;
+use App\Modules\Shared\Domain\Entities\UserPermission;
+use App\Modules\Shared\Infrastructure\Exceptions\AccessDeniedException;
 
 class RemoveClientUseCase
 {
@@ -13,9 +15,11 @@ class RemoveClientUseCase
     )
     {
     }
-    public function execute(int $id): bool
+    public function execute(int $id, UserPermission $permissions): bool
     {
-        //TODO Проверка, можем ли удалить
+        if (!$permissions->can('auth.buyer.delete')) throw new AccessDeniedException();
+
+        //Проверка, можем ли удалить
 
         return $this->clientRepository->delete($id);
     }
