+++ app/Modules/Auth/Application/Actions/Freelance/RemoveFreelanceUseCase.php
@@ -3,6 +3,8 @@
 namespace App\Modules\Auth\Application\Actions\Freelance;
 
 use App\Modules\Auth\Application\Interfaces\FreelanceRepositoryInterface;
+use App\Modules\Shared\Domain\Entities\UserPermission;
+use App\Modules\Shared\Infrastructure\Exceptions\AccessDeniedException;
 
 class RemoveFreelanceUseCase
 {
@@ -11,9 +13,10 @@ class RemoveFreelanceUseCase
     )
     {
     }
-    public function execute(int $id): bool
+    public function execute(int $id, UserPermission $permissions): bool
     {
-        //TODO Проверка, можем ли удалить
+        if (!$permissions->can('auth.employee.delete')) throw new AccessDeniedException();
+        //Проверка, можем ли удалить
 
         return $this->freelanceRepository->delete($id);
     }
